requirements.txt
dice~=3.1.2
discord.py~=1.7.1
petname~=2.6
emoji~=1.2.0
pydealer~=1.4.0
